other/panda365/panda365/requirements/prod.txt
Faker==0.7.17
Flask-BabelEx==0.9.3
Werkzeug==0.12.2
alembic==0.9.2
amqp==2.2.1
arrow==0.10.0
babel==2.4.0
bcrypt==3.1.3
boto==2.48.0
celery==4.0.2
datadog==0.16.0
decorator==4.1.1
factory-boy==2.8.1
flask-admin==1.5.0
flask-cors==3.0.2
flask-limiter==0.9.4
flask-migrate==2.0.4
flask-security-fork==2.0.1
flask-sqlalchemy==2.2
flask==0.12.2
furl==1.0.0
git+https://github.com/mobolic/facebook-sdk.git@ea4e43c#egg=facebook-sdk
markdown==2.6.8
marshmallow-sqlalchemy==0.13.1
marshmallow==2.13.5
pillow==4.2.0
psycopg2==2.7.1
pyjwt==1.5.0
raven==6.1.0
redis==2.10.5
requests==2.18.1
setuptools==36.2.0
simplejson==3.11.1
sqlalchemy-utils==0.32.14
sqlalchemy==1.1.11
webargs==1.8.0
gunicorn==19.7.1
